requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy>=2.0.40
alembic==1.14.0
# psycopg2-binary==2.9.10  # Only needed for PostgreSQL in production (Render injects this)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
pydantic-settings==2.6.1
jinja2==3.1.4
boto3==1.35.76
python-dotenv==1.0.1
aiofiles==24.1.0
